// src/api/bulletin/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface BulletinApi {
	bulletinList: (request: Api.BulletinBulletinListRequest) => Promise<BulletinListResponse>
}

export type BulletinListResponse = BulletinList200Response | BulletinListDefaultResponse

export interface BulletinList200Response {
	status: 200
	body: Api.BulletinBulletinListResponse
	headers?: never
}

export interface BulletinListDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}